Added a README in english.
We're going to speak english here. Yeah, that was for a french team. But
it is now for a **world wide international team**!

Just kidding, but a serious open source project should be in english, so
that coders all over the world can profit from it. At least, the one
speaking english (but you have to speak english to code, don't you?).