Added stuff that all C project have.
Things like the main function, header defines...

Also, this is a little messy for now. I've been waiting too long to find
a clean way to do right at the beginning. I'll write, then clean, then
write, and so on. You can't only think, you have to act.

So, for now, I'm going to rush in some code. The redaction effort will
come later (only the redaction effort, the explanation will be there
with the code, just not well written).